pyproject.toml
[project]
name = "cloudflare-manager"
version = "0.2.0"
description = "Multi-tenant Telegram bot for managing Cloudflare DNS records and tunnels with role-based access control"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "cloudflare>=4.3.1",
    "python-dotenv>=1.0.0",
    "python-telegram-bot[http2]>=22.3",
    "sqlalchemy[asyncio]>=2.0.43",
]

[tool.ruff]
line-length = 120
indent-width = 4
respect-gitignore = true

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = "dynamic"
